+++ lib/src/http/request/request.dart
@@ -8,6 +8,10 @@ import 'package:mime/mime.dart';
 import 'package:http_server/http_server.dart';
 import 'package:logging/logging.dart';
 
+import 'package:jaguar/src/http/session/session.dart';
+
+part 'form_field.dart';
+
 /// Splits given path to composing segments
 List<String> splitPathToSegments(final String paths) {
   final List<String> segments = paths.split(new RegExp('/+'));
@@ -20,12 +24,103 @@ List<String> splitPathToSegments(final String paths) {
 }
 
 /// [Request] contains information about HTTP request
-class Request {
+abstract class Request {
+  factory Request(
+          HttpRequest request, SessionManager sessionManager, Logger log) =>
+      new _Request(request, sessionManager, log);
+
+  /// The URI for the request
+  ///
+  /// This provides access to the path and query string for the request.
+  Uri get uri;
+
+  /// The requested URI for the request
+  ///
+  /// The returned URI is reconstructed by using http-header fields, to access
+  /// otherwise lost information, e.g. host and scheme.
+  ///
+  /// To reconstruct the scheme, first 'X-Forwarded-Proto' is checked, and then
+  /// falling back to server type.
+  ///
+  /// To reconstruct the host, first 'X-Forwarded-Host' is checked, then 'Host'
+  /// and finally calling back to server.
+  Uri get requestedUri;
+
+  /// The method, such as 'GET' or 'POST', for the request.
+  String get method;
+
+  /// The request headers.
+  HttpHeaders get headers;
+
+  /// The cookies in the request, from the Cookie headers.
+  List<Cookie> get cookies;
+
+  /// The session for the given request
+  Future<Session> get session;
+
+  /// Connection information of the request
+  HttpConnectionInfo get connectionInfo;
+
+  /// Logger to log data during request
+  Logger get log;
+
+  /// The client certificate of the client making the request
+  ///
+  /// This value is null if the connection is not a secure TLS or SSL connection,
+  /// or if the server does not request a client certificate, or if the client
+  /// does not provide one.
+  X509Certificate get certificate;
+
+  /// The content length of the request body
+  ///
+  /// If the size of the request body is not known in advance,
+  /// this value is -1.
+  int get contentLength;
+
+  /// The persistent connection state signaled by the client
+  bool get persistentConnection;
+
+  /// The HTTP protocol version used in the request, either "1.0" or "1.1".
+  String get protocolVersion;
+
+  Future<WebSocket> get upgradeToWebSocket;
+
+  /// Returns raw body of HTTP request
+  Future<List<int>> get body;
+
+  /// Returns the body of HTTP request as Stream
+  Future<Stream<List<int>>> get bodyAsStream;
+
+  /// Returns body as text
+  Future<String> bodyAsText([Encoding encoding = UTF8]);
+
+  /// Decodes JSON body of the request
+  Future<dynamic> bodyAsJson({Encoding encoding: UTF8});
+
+  /// Decodes JSON body of the request as [Map]
+  Future<Map> bodyAsJsonMap({Encoding encoding: UTF8});
+
+  /// Decodes JSON body of the request as [List]
+  Future<List> bodyAsJsonList({Encoding encoding: UTF8});
+
+  /// Decodes url-encoded form from the body and returns Map<String, String>
+  Future<Map<String, String>> bodyAsUrlEncodedForm({Encoding encoding: UTF8});
+
+  /// Decodes `multipart/form-data` body
+  Future<Map<String, FormField>> bodyAsFormData({Encoding encoding: UTF8});
+}
+
+/// [Request] contains information about HTTP request
+class _Request implements Request {
   final HttpRequest _request;
 
   final Logger log;
 
-  Request(this._request, this.log);
+  final SessionManager _sessionManager;
+
+  Session _session;
+
+  _Request(this._request, this._sessionManager, this.log);
 
   /// The client certificate of the client making the request.
   X509Certificate get certificate => _request.certificate;
@@ -55,10 +150,19 @@ class Request {
   Uri get requestedUri => _request.requestedUri;
 
   /// The session for the given request.
-  HttpSession get session => _request.session;
+  Future<Session> get session async {
+    if (_session == null) {
+      _session = await _sessionManager.parse(this);
+    }
+    return this._session;
+  }
 
   /// The URI for the request.
   Uri get uri => _request.uri;
+
+  /// Upgrades the request to websocket request
+  Future<WebSocket> get upgradeToWebSocket =>
+      WebSocketTransformer.upgrade(_request);
 
   /// Private cache for request body
   List<int> _body;
@@ -84,10 +188,6 @@ class Request {
   Future<String> bodyAsText([Encoding encoding = UTF8]) async {
     return encoding.decode(await body);
   }
-
-  /// Upgrades the request to websocket request
-  Future<WebSocket> get upgradeToWebSocket =>
-      WebSocketTransformer.upgrade(_request);
 
   /// Decodes JSON body of the request
   Future<dynamic> bodyAsJson({Encoding encoding: UTF8}) async {
@@ -176,117 +276,5 @@ class Request {
     }
 
     return ret;
-  }
-}
-
-/// Abstract class for fields on `multipart/form-data`
-abstract class FormField<T> {
-  /// Name of the field
-  String get name;
-
-  /// Value of the field
-  T get value;
-
-  /// content-type of the field
-  ContentType get contentType;
-
-  /// Hash code
-  int get hashCode => name.hashCode;
-}
-
-/// String field in the `multipart/form-data`
-///
-/// Contains `String` value
-class StringFormField implements FormField {
-  /// Name of the field
-  final String name;
-
-  /// Value of the field
-  final String value;
-
-  /// content-type of the field
-  final ContentType contentType;
-
-  StringFormField(this.name, this.value, {this.contentType});
-
-  bool operator ==(other) {
-    if (other is! StringFormField) {
-      return false;
-    }
-
-    return _equality(other as StringFormField);
-  }
-
-  bool _equality(StringFormField other) {
-    return name == other.name &&
-        value == other.value &&
-        contentType.mimeType == other.contentType.mimeType &&
-        contentType.charset == other.contentType.charset;
-  }
-
-  int get hashCode => name.hashCode;
-
-  String toString() {
-    return "StringFormField('$name', '$value', '$contentType')";
-  }
-}
-
-class TextFileFormField implements FormField {
-  /// Name of the field
-  final String name;
-
-  /// Value of the field
-  ///
-  /// Returns a stream of text
-  final Stream<String> value;
-
-  /// content-type of the field
-  final ContentType contentType;
-
-  /// File of the file field
-  final String filename;
-
-  TextFileFormField(this.name, this.value, {this.contentType, this.filename});
-
-  bool operator ==(other) {
-    return name == other.name &&
-        contentType.mimeType == other.contentType.mimeType &&
-        contentType.charset == other.contentType.charset &&
-        filename == other.filename;
-  }
-
-  int get hashCode => name.hashCode;
-
-  String toString() {
-    return "FileFormField('$name', '$contentType', '$filename')";
-  }
-}
-
-class BinaryFileFormField implements FormField {
-  /// Name of the field
-  final String name;
-
-  /// Value of the field
-  final Stream<List<int>> value;
-
-  /// content-type of the field
-  final ContentType contentType;
-
-  /// File of the file field
-  final String filename;
-
-  BinaryFileFormField(this.name, this.value, {this.contentType, this.filename});
-
-  bool operator ==(other) {
-    return name == other.name &&
-        contentType.mimeType == other.contentType.mimeType &&
-        contentType.charset == other.contentType.charset &&
-        filename == other.filename;
-  }
-
-  int get hashCode => name.hashCode;
-
-  String toString() {
-    return "FileFormField('$name', '$contentType', '$filename')";
   }
 }